Chamadas State Holder -> Reutilização mais simples -> Reutilização mais refinada e com mais controle e flexibilidade.
Conceitos de stateful & stateless
