+++ app/src/main/java/com/example/aludelivery/ui/activities/MainActivity.kt
@@ -25,6 +25,7 @@ import com.example.aludelivery.sampledata.sampleCandies
 import com.example.aludelivery.sampledata.sampleDrinks
 import com.example.aludelivery.sampledata.sampleProducts
 import com.example.aludelivery.sampledata.sampleSections
+import com.example.aludelivery.ui.screens.HomeScreen
 import com.example.aludelivery.ui.screens.HomeScreenContent
 import com.example.aludelivery.ui.screens.HomeScreenUiState
 import com.example.aludelivery.ui.theme.AluDeliveryTheme
@@ -45,45 +46,7 @@ class MainActivity : ComponentActivity() {
                 )
             }) {
                 val products = dao.products()
-                val sections = mapOf(
-                    "Todos produtos" to products,
-                    "Promoções" to sampleDrinks + sampleCandies,
-                    "Doces" to sampleCandies,
-                    "Bebidas" to sampleDrinks
-                )
-                var text by remember {
-                    mutableStateOf("")
-                }
-
-                fun containsNameOrDescription() = { product: Product ->
-                    product.name.contains(
-                        text,
-                        ignoreCase = true,
-                    ) ||
-                            product.description?.contains(
-                                text,
-                                ignoreCase = true,
-                            ) ?: false
-                }
-
-                val searchedProducts = remember(text, products) {
-                    if (text.isNotBlank()) {
-                        sampleProducts.filter(containsNameOrDescription()) +
-                                products.filter(containsNameOrDescription())
-                    } else emptyList()
-                }
-
-                val state = remember(products, text) {
-                    HomeScreenUiState(
-                        sections = sections,
-                        searchedProducts = searchedProducts,
-                        searchText = text,
-                        onSearchChange = {
-                            text = it
-                        }
-                    )
-                }
-                HomeScreenContent(state = state)
+                HomeScreen(products = products)
             }
         }
     }
