+++ app/src/main/java/com/example/aludelivery/ui/screens/HomeScreens.kt
@@ -10,10 +10,17 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.aludelivery.model.Product
+import com.example.aludelivery.sampledata.sampleCandies
+import com.example.aludelivery.sampledata.sampleDrinks
+import com.example.aludelivery.sampledata.sampleProducts
 import com.example.aludelivery.sampledata.sampleSections
 import com.example.aludelivery.ui.components.CardProductItem
 import com.example.aludelivery.ui.components.ProductSection
@@ -30,6 +37,50 @@ class HomeScreenUiState(
     fun isShowSections(): Boolean {
         return searchText.isBlank()
     }
+}
+
+@Composable
+fun HomeScreen(products: List<Product>) {
+    val sections = mapOf(
+        "Todos produtos" to products,
+        "Promoções" to sampleDrinks + sampleCandies,
+        "Doces" to sampleCandies,
+        "Bebidas" to sampleDrinks
+    )
+
+    var text by remember {
+        mutableStateOf("")
+    }
+
+    fun containsNameOrDescription() = { product: Product ->
+        product.name.contains(
+            text,
+            ignoreCase = true,
+        ) ||
+                product.description?.contains(
+                    text,
+                    ignoreCase = true,
+                ) ?: false
+    }
+
+    val searchedProducts = remember(text, products) {
+        if (text.isNotBlank()) {
+            sampleProducts.filter(containsNameOrDescription()) +
+                    products.filter(containsNameOrDescription())
+        } else emptyList()
+    }
+
+    val state = remember(products, text) {
+        HomeScreenUiState(
+            sections = sections,
+            searchedProducts = searchedProducts,
+            searchText = text,
+            onSearchChange = {
+                text = it
+            }
+        )
+    }
+    HomeScreenContent(state = state)
 }
 
 @Composable
